requirements.txt
asgiref==3.2.9
backcall==0.2.0
certifi==2020.4.5.2
chardet==3.0.4
decorator==4.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-extensions==2.2.9
django-filter==2.3.0
django-input-mask==2.0.2
django-localflavor==3.0.1
django-nose==1.4.6
djangorestframework==3.11.0
drf-writable-nested==0.6.0
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
psycopg2==2.8.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-decouple==3.3
python-stdnum==1.13
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
